refactor(hooks/incident): break `useCurrentIncident` into `hooks/incident` submodule
